+++ src/tools/ufo2map/common/polylib.c
@@ -253,11 +253,11 @@ CopyWinding
 */
 winding_t	*CopyWinding (winding_t *w)
 {
-	int			size;
+	ptrdiff_t	size;
 	winding_t	*c;
 
 	c = AllocWinding (w->numpoints);
-	size = (int)((winding_t *)0)->p[w->numpoints];
+	size = (ptrdiff_t)((winding_t *)0)->p[w->numpoints];
 	memcpy (c, w, size);
 	return c;
 }
